+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Support content-type aware combinators and `Accept`/`Content-type` headers
 * Added a lot of tests
 * Support multiple concurrent threads
+* Use `ServantError` to report Errors instead of `String`
 
 0.2.2
 -----